+++ bytecode/flags.py
@@ -14,19 +14,21 @@ class CoFlags(IntEnum):
     furthermore would be version dependent.
 
     """
-    CO_OPTIMIZED             = 0x0001
-    CO_NEWLOCALS             = 0x0002
-    CO_VARARGS               = 0x0004
-    CO_VARKEYWORDS           = 0x0008
-    CO_NESTED                = 0x0010
-    CO_GENERATOR             = 0x0020
-    CO_NOFREE                = 0x0040
+    CO_OPTIMIZED             = 0x00001  # noqa
+    CO_NEWLOCALS             = 0x00002  # noqa
+    CO_VARARGS               = 0x00004  # noqa
+    CO_VARKEYWORDS           = 0x00008  # noqa
+    CO_NESTED                = 0x00010  # noqa
+    CO_GENERATOR             = 0x00020  # noqa
+    CO_NOFREE                = 0x00040  # noqa
     # New in Python 3.5
-    CO_COROUTINE             = 0x0080
-    CO_ITERABLE_COROUTINE    = 0x0100
-    #♣ New in Python 3.6
-    CO_ASYNC_GENERATOR       = 0x0200
-    CO_FUTURE_GENERATOR_STOP = 0x80000
+    CO_COROUTINE             = 0x00080  # noqa
+    CO_ITERABLE_COROUTINE    = 0x00100  # noqa
+    # New in Python 3.6
+    CO_ASYNC_GENERATOR       = 0x00200  # noqa
+
+    # __future__ flags
+    CO_FUTURE_GENERATOR_STOP = 0x80000  # noqa
 
 
 _CAN_DEDUCE_FROM_CODE = ('optimized', 'generator', 'nofree', 'coroutine')
@@ -36,8 +38,8 @@ class Flags:
     """Object representing the flags set for the bytecode.
 
     Flags can be bools indicating their state, or for some of them None
-    indicating that the value should be inferred from the bytecode object
-    passed when converting to an integer. If no bytecode is provided the
+    indicating that the value should be inferred from the ConcreteBytecode
+    object passed when converting to an integer. If no bytecode is provided the
     default values are used.
 
     The flags whose natural values can be inferred are:
@@ -48,13 +50,10 @@ class Flags:
 
     """
 
-    def __init__(self, int_or_flags=None):
+    def __init__(self, int_or_flags=0):
 
-        self._defaults = defaultdict(bool)
         self._forced = {}
 
-        if int_or_flags is None:
-            return
         if isinstance(int_or_flags, Flags):
             self._defaults = int_or_flags._defaults.copy()
             self._forced = int_or_flags._forced.copy()
@@ -80,19 +79,26 @@ class Flags:
 
         flags = defaultdict(bool)
         flags.update(self._forced)
-        if bytecode:
+        print(bytecode)
+        if bytecode is not None:
+            if not isinstance(bytecode, _bytecode.ConcreteBytecode):
+                msg = 'Expected a ConcreteBytecode instance not %s'
+                raise ValueError(msg % bytecode)
             instr_names = {i.name for i in bytecode
                            if not isinstance(i, _bytecode.SetLineno)}
+
+        print(self._defaults)
         for k, v in self._defaults.items():
             if k not in flags:
-                if k in _CAN_DEDUCE_FROM_CODE and bytecode:
+                print(k in _CAN_DEDUCE_FROM_CODE, bytecode is not None)
+                if k in _CAN_DEDUCE_FROM_CODE and bytecode is not None:
+                    print('y')
                     if k == 'optimized':
                         flags[k] = not (instr_names & {'STORE_NAME',
                                                        'LOAD_NAME',
                                                        'DELETE_NAME'})
                     elif k == 'generator':
-                        if not any((self.async_generator, self.coroutine,
-                                    self.iterable_coroutine)):
+                        if not self.async_generator:
                             flags[k] = bool(instr_names & {'YIELD_VALUE',
                                                            'YIELD_FROM'})
                         else:
@@ -104,7 +110,8 @@ class Flags:
                                                        'DELETE_DEREF',
                                                        'LOAD_CLASSDEREF'})
                     else:
-                        if not self.iterable_coroutine:
+                        if not any((self.async_generator,
+                                    self.iterable_coroutine)):
                             flags[k] = bool(instr_names & {'GET_AWAITABLE',
                                                            'GET_AITER',
                                                            'GET_ANEXT',
@@ -138,10 +145,11 @@ class Flags:
 
     @optimized.setter
     def optimized(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
+        print(value)
         if value is None:
             del self._forced['optimized']
+        elif not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean or None')
         else:
             self._forced['optimized'] = value
 
@@ -202,14 +210,14 @@ class Flags:
         """The compiled code block is a generator code block.
 
         """
-        return self._forced.get('generator', self._defaults['generator'])
+        return self._forced.get('generator')
 
     @generator.setter
     def generator(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
         if value is None:
             del self._forced['generator']
+        elif not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean or None')
         else:
             self._forced['generator'] = value
 
@@ -222,10 +230,10 @@ class Flags:
 
     @nofree.setter
     def nofree(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
         if value is None:
             del self._forced['nofree']
+        elif not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean or None')
         else:
             self._forced['nofree'] = value
 
@@ -240,10 +248,10 @@ class Flags:
 
     @coroutine.setter
     def coroutine(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
         if value is None:
             del self._forced['coroutine']
+        elif not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean or None')
         else:
             self._forced['coroutine'] = value
 
@@ -256,16 +264,14 @@ class Flags:
         New in Python version 3.5
 
         """
-        return self._forced.get('iterable_coroutine')
+        return self._forced.get('iterable_coroutine',
+                                self._defaults['iterable_coroutine'])
 
     @iterable_coroutine.setter
     def iterable_coroutine(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
-        if value is None:
-            del self._forced['iterable_coroutine']
-        else:
-            self._forced['iterable_coroutine'] = value
+        if not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean')
+        self._forced['iterable_coroutine'] = value
 
     @property
     def async_generator(self):
@@ -274,16 +280,14 @@ class Flags:
         New in Python version 3.6
 
         """
-        return self._forced.get('async_generator')
+        return self._forced.get('async_generator',
+                                self._defaults['async_generator'])
 
     @async_generator.setter
     def async_generator(self, value):
-        if not isinstance(value, bool) or value is None:
-            raise ValueError('Flag value should be a boolean or None')
-        if value is None:
-            del self._forced['async_generator']
-        else:
-            self._forced['async_generator'] = value
+        if not isinstance(value, bool):
+            raise ValueError('Flag value should be a boolean')
+        self._forced['async_generator'] = value
 
     @property
     def future_generator_stop(self):